Daten von WeeWX annehmen

+++ src/main/java/org/voegtle/weatherstation/server/RapidService.kt
@@ -17,7 +17,7 @@ import java.util.logging.Logger
               @RequestParam PASSWORD: String,
               @RequestParam dateutc: String,
               @RequestParam temp: Float,
-              @RequestParam humidity: Int,
+              @RequestParam humidity: Float,
               @RequestParam barometer: Float,
               @RequestParam dailyrain: Float,
               @RequestParam rain: Float,
